@@ -18,16 +18,18 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-project.ext.set('VERSION_CODE', 2)
-project.ext.set('VERSION_NAME', '1.1')
+project.ext.set('VERSION_CODE', 3)
+project.ext.set('VERSION_NAME', '1.2')
 
 buildscript {
     repositories {
         mavenCentral()
+        google()
+        jcenter()
     }
 
     dependencies {
-        classpath 'com.android.tools.build:gradle:1.2.3'
+        classpath 'com.android.tools.build:gradle:4.1.1'
     }
 }
 
@@ -35,28 +37,33 @@ apply plugin: 'com.android.library'
 
 repositories {
     mavenCentral()
+    google()
+    jcenter()
 }
 
 dependencies {
-    compile "com.android.support:support-annotations:22.2.0"
-    androidTestCompile 'junit:junit:4.12'
-    androidTestCompile('com.android.support.test:runner:0.2') {
-        exclude group: 'junit'
-    }
+    implementation "androidx.annotation:annotation:1.1.0"
+    androidTestImplementation 'junit:junit:4.12'
+    androidTestImplementation 'androidx.test:runner:1.1.0'
 }
 
 
 android {
-    compileSdkVersion 22
-    buildToolsVersion "22.0.1"
+    compileSdkVersion 30
 
     defaultConfig {
         versionCode project.VERSION_CODE
         versionName project.VERSION_NAME
-        minSdkVersion 8
+        minSdkVersion 14
         targetSdkVersion 22
 
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        consumerProguardFile "consumer-rules.pro"
+
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+    }
+
+    buildFeatures {
+        buildConfig false
     }
 }
 
@@ -65,4 +72,3 @@ if (project.hasProperty('badooMavenScript')) {
 } else {
     apply from: 'mvn-local-support.gradle'
 }
-
